move all test vectors to zebra-test
